+++ CHANGELOG.md
@@ -1,4 +1,6 @@
+ * Added a new method to get elementary stream info from IVideoInput
  * Video output now handles input images with arbitrary underlying raster type
+ * Implemented ST 0601 tags 90-93
 
 ### June 6, 2018 version 1.2.0
  * Added the ability to write to a UDP stream